backend/requirements.txt
# FastAPI + Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
websockets==14.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Redis
redis==5.2.1

# Celery
celery[redis]==5.4.0

# Auth
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# HTTP Client
httpx==0.28.1

# Blockchain
web3==7.6.0
bip-utils==2.9.3
eth-account==0.13.4

# QR Code
qrcode[pil]==8.0

# Rate Limiting
slowapi==0.1.9

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
aiosqlite==0.20.0

# Crypto
cryptography==44.0.0

# Date handling
python-dateutil==2.9.0

# Email
resend==2.5.0
